feat: Update transsaction detail interface
