+++ src/stores/finance.js
@@ -5,9 +5,16 @@ export const useFinanceStore = defineStore('finance', {
     state: () => ({}),
     getters: {},
     actions: {
-        async createTransaction({payload, quotationId, projectDealUid}) {
+        async createTransaction({payload, projectDealUid}) {
             try {
-                return await axios.post(`/finance/transaction/${quotationId}/${projectDealUid}`, payload);
+                return await axios.post(`/finance/transaction/${projectDealUid}`, payload);
+            } catch (error) {
+                return error;
+            }
+        },
+        async generateBillInvoice(payload, projectDealUid) {
+            try {
+                return await axios.post(`/finance/${projectDealUid}/billInvoice`, payload);
             } catch (error) {
                 return error;
             }
